telecom_erp_api/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==4.2.23
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
